requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
aiohttp>=3.9.0
feedparser>=6.0.0
praw>=7.7.0
aiosqlite>=0.19.0
pydantic>=2.5.0
pyyaml>=6.0.0
